Change to GFDL for future met
MACA does not include longwave radiation or air pressure, so we can't use it for ED.  Switching to GFLD ensemble members instead.  This will still let us get a couple variants of climate change to test and has the added bonus of already being 3-hourly resolution.  This means if we aggregate NLDAS (as the training dataset) to that resolution (script #1), then we can get rid of the Temporal Downscaling step and just bias-correct.  Hopefully it works...