Added the same conversion scheme as ED2 and a section to check the base weather files
I've now started looking at the weather data at all steps of the process and have added a section that directly compares the ED2 output and all the ncdf (or h5) files that occur immediately before and after ED2 is run to understand where the mismatch occurs. CUrrently it seem to occur in the shell script in a way I don't understand.