Added points of comparison for every step of weather file
Comparing every weather file it's clear that a discrepancy emerges between the 0_setup workflow and the 1_runs workflow. Before I even pull out into the 2_analysis, the underlying weather is different. I can now show this a few ways and think it proves my point something is going on in the blackbox of the shell script