Attempt at visualizing multiple comparison
This is honestly driving me crazy. Any premade solutions don't work with the way we have our models run and data organized so I've been trying to write out a loop that will allow us to assign groups. I've gotten close but the final steps of how to identify the groups and paste on labels is driving me up a wall. I've spent more than 4 hours today trying to solve this and so I think it's going to be punted until later in the graphics cleaning process because I'm struggling and need to make progress elsewhere