Trying to work with Management and failing
When I try to include management in the interaction it alwasy excludes some level. I can't get any interactions that involve the crash lag (-5) or the management condition "None" both of these are the first level factors in their groups and are being excluded. I am including the -1 to remove one of them being label as intercept but even then there aren't any interactions with either. I'm not sure why but I need all the interactions so this isn't functional until I solve it