Added generic copies for relative weather and structure
They will need to be changed and updated but I made copies of my original analysis so we now have a version that looks at relative weather and structural variables. They still suffer from the issue of not reporting effects for lag -5 and None management. I'm really struggling to fix that and might need to leave that on the burner for when Christy gets back.