Finally understand the centering loop
I was able to understand and implement the centering loop. I understand what it is technically doing (relativizing the values by how they differ from the mean of the previous period) but I couldn't articulate well why we do this. I guess we want to investigate how the variables differ within that 5 year period