Added new version where whether the style crashes is a predictor
If I only use management as a predictor I'm not really comparing which crashed and which didn't. This way I am investigating a difference in the structure before crash for those that did crash vs. those that didn't. Maybe this is redudant to an analysis that only uses data where a crash occurs and investigates the differences beforehand but I think there is something more here. At the very least, I know comparing across management without knowing which crashed and which doesn't seem useful