Now I capture duplicates
I now capture all situations where one year can be a different lag for different crash years. We now label crash years and I make a quick dataframe of the duplicates that gets rbind in at the end. I am still having an issue with the situtaitons where we try to group managements when only one crashes. That's next