Removed old figure and added new table
Removed a loop that attempt to add significance markers between management types using the Tukey. It was a mess that I never got working. In its place I added a loop that does a basic anova on each strucutrual variable at three time periods to with mangement and rcp as the predictors