Still trying to update the group SEA calculation
I've been able to get the ind.lag.group toconsistent be the appropriate 94 crashes but now am only getting about 250 for group instead of the 376 that should be occuring. It seems like some of the ind values may also be shifted a year off but I can't tell why. This has really ballooned into something tricky